Improve Welsh localization support
This commit improves Welsh-language localization in the shared styles
for PPD, std-reports and the landing site. In particular, we now
localize the text in the header menu and the footer. `LrCommonConfig`
has been simplified by moving some standard messages into the locale
message catalogues.

A standard language-switcher partial is also provided, for those sites
that can make use of it.